chore: rename privy evm wallet provider
